Added PanoramaPublicSymlinkTest
